Add Set/GetOmnisharpPort commands for C# completer
These commands allow setting the port it uses to communicate with the
Omnisharp server. This allows sharing of the Omnisharp server by
retrieving the port and using it in other clients, or get the port
another client is using and using it here.

+++ ycmd/completers/cs/cs_completer.py
@@ -186,7 +186,16 @@ class CsharpCompleter( Completer ):
       'ServerIsReady'                    : ( lambda self, request_data, args:
          self._SolutionSubcommand( request_data,
                                    method = 'ServerIsReady',
-                                   no_request_data = True ) )
+                                   no_request_data = True ) ),
+      'SetOmnisharpPort'                    : ( lambda self, request_data, args:
+         self._SolutionSubcommand( request_data,
+                                   method = '_SetOmnisharpPort',
+                                   port = args[ 0 ],
+                                   no_request_data = True ) ),
+      'GetOmnisharpPort'                    : ( lambda self, request_data, args:
+         self._SolutionSubcommand( request_data,
+                                   method = '_GetOmnisharpPort',
+                                   no_request_data = True ) ),
     }
 
 
@@ -314,6 +323,7 @@ class CsharpSolutionCompleter:
     self._omnisharp_phandle = None
     self._desired_omnisharp_port = desired_omnisharp_port
     self._server_state_lock = threading.RLock()
+    self._external_omnisharp = False
 
 
   def CodeCheck( self, request_data ):
@@ -365,22 +375,24 @@ class CsharpSolutionCompleter:
         with open( self._filename_stdout, 'w' ) as fstdout:
           self._omnisharp_phandle = utils.SafePopen(
               command, stdout = fstdout, stderr = fstderr )
+          self._external_omnisharp = False
 
       self._solution_path = path_to_solutionfile
+
 
 
   def _StopServer( self ):
     """ Stop the OmniSharp server using a lock. """
     with self._server_state_lock:
-      if not self.ServerIsRunning():
-        return
+      #if not self.ServerIsRunning():
+      #  return
 
       self._logger.info( 'Stopping OmniSharp server' )
 
       self._TryToStopServer()
 
       # Kill it if it's still up
-      if self.ServerIsRunning():
+      if self.ServerIsRunning() and not self.ServerIsExternal():
         self._logger.info( 'Killing OmniSharp server' )
         self._omnisharp_phandle.kill()
 
@@ -525,9 +537,17 @@ class CsharpSolutionCompleter:
     return parameters
 
 
+
+  def ServerIsExternal( self ):
+    return self._external_omnisharp
+
+
   def ServerIsRunning( self ):
     """ Check if our OmniSharp server is running (process is up)."""
-    return utils.ProcessIsRunning( self._omnisharp_phandle )
+    if self.ServerIsExternal():
+      return self._omnisharp_port is not None
+    else:
+      return utils.ProcessIsRunning( self._omnisharp_phandle )
 
 
   def ServerIsHealthy( self ):
@@ -561,6 +581,18 @@ class CsharpSolutionCompleter:
     # We cannot use 127.0.0.1 like we do in other places because OmniSharp
     # server only listens on localhost.
     return 'http://localhost:' + str( self._omnisharp_port )
+
+
+  def _GetOmnisharpPort( self ):
+    return responses.BuildDisplayMessageResponse( self._omnisharp_port )
+
+
+  def _SetOmnisharpPort( self, port ):
+    if self.ServerIsRunning():
+        self.StopServer()
+
+    self._omnisharp_port = port
+    self._external_omnisharp = True
 
 
   def _GetResponse( self, handler, parameters = {}, timeout = None ):
